DatagovukPlugin: add before_index method
solr6+ limits fields of type StrField to a length of ~32k, but ckan's
schema.xml has a catchall field which will add unmentioned fields as
StrFields. this means we can only afford to have strings greater than 32k
under keys we're pretty sure have TextFields for. take this list of
"safe" field name patterns from the ckan28 one as @
https://github.com/alphagov/govuk-puppet/blob/a927fcdf8210bfb4f810ca619a0c4df9fcde425a/modules/govuk_solr6/files/ckan28.schema.xml

additionally, when we truncate strings, truncate them to under half the
32k limit as it's possible we may have >1 such string under any one field.
this is obviously not a complete solution for this, but works for now. in
future we may have to do something clever involving performing the json-
squashing of complex values ourselves so we can know what the final value
length is going to be.